partial index check to test utils
